Move DownloadGitHubRawFile function to VersionIdCell and add extension to downloaded file
