+++ src/components/CorpusMetadata/TableComponent/TableBody/MoreCell.js
@@ -2,45 +2,6 @@ import { useContext } from "react";
 import { Checkbox, TableCell, Tooltip } from "@mui/material";
 import { Context } from "../../../../App";
 
-// download raw github file
-export function downloadGitHubRawFile(row) {
-  const outputFilename = `${row?.version_uri}.txt`;
-
-  fetch(row?.release_version?.url)
-    .then((response) => {
-      if (!response.ok) {
-        throw new Error("Failed to download file.");
-      }
-      return response.blob();
-    })
-    .then((blob) => {
-      // Create a temporary anchor element
-      const anchor = document.createElement("a");
-      anchor.style.display = "none";
-      document.body.appendChild(anchor);
-
-      // Create a URL object from the blob
-      const url = window.URL.createObjectURL(blob);
-
-      // Set the anchor's href to the URL
-      anchor.href = url;
-
-      // Set the anchor's download attribute and filename
-      anchor.download = outputFilename;
-
-      // Trigger a click event on the anchor element to start the download
-      anchor.click();
-
-      // Clean up by revoking the URL object
-      window.URL.revokeObjectURL(url);
-
-      // Remove the temporary anchor element from the document
-      document.body.removeChild(anchor);
-    })
-    .catch((error) => {
-      console.error("Error:", error.message);
-    });
-}
 
 const MoreCell = ({ classes, row }) => {
   const { checkedBooks, setCheckedBooks } = useContext(Context);
